# hivepulse-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Open-source uptime monitoring platform",
        "title": "HivePulse API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/monitors": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitors"
                ],
                "summary": "List monitors",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitors"
                ],
                "summary": "Create monitor",
                "parameters": [
                    {
                        "description": "Monitor payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.monitorWriteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/monitors/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitors"
                ],
                "summary": "Get monitor by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Monitor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "monitors"
                ],
                "summary": "Update monitor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Monitor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Monitor payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.monitorWriteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "monitors"
                ],
                "summary": "Delete monitor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Monitor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/monitors/{id}/heartbeats": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitors"
                ],
                "summary": "List heartbeats for a monitor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Monitor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 48,
                        "description": "Number of heartbeats",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/monitors/{id}/stats": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitors"
                ],
                "summary": "Get aggregated stats for a monitor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Monitor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Time range: 1h, 3h, 6h, 24h, 48h, 7d, 15d, 30d, or 90d",
                        "name": "range",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.StatsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.DownPeriod": {
            "type": "object",
            "properties": {
                "resolved_at": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                }
            }
        },
...
        "domain.StatsBucket": {
            "type": "object",
            "properties": {
                "avg_ping_ms": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                },
                "up_count": {
                    "type": "integer"
                }
            }
        },
        "domain.StatsResponse": {
            "type": "object",
            "properties": {
                "avg_ping_ms": {
                    "type": "integer"
                },
                "buckets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.StatsBucket"
                    }
                },
                "down_periods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.DownPeriod"
                    }
                },
                "uptime_pct": {
                    "type": "number"
                }
            }
        },
...
        "handler.monitorWriteRequest": {
            "type": "object",
            "required": [
                "check_type",
                "interval",
                "name"
            ],
            "properties": {
                "check_type": {
                    "type": "string"
                },
                "dns_host": {
                    "type": "string"
                },
                "dns_server": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "expected_status": {
                    "type": "integer"
                },
                "expected_value": {
                    "type": "string"
                },
                "follow_redirects": {
                    "type": "boolean"
                },
                "host": {
                    "type": "string"
                },
                "interval": {
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "packet_count": {
                    "type": "integer"
                },
                "ping_host": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "record_type": {
                    "type": "string"
                },
                "request_body": {
                    "type": "string"
                },
                "request_headers": {
                    "type": "string"
                },
                "retries": {
                    "type": "integer"
                },
                "retry_interval": {
                    "type": "integer"
                },
                "skip_tls_verify": {
                    "type": "boolean"
                },
                "timeout": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}